address a bunch of potential issues pointed out by CppCheck

+++ src/Engine/InteractiveSurface.h
@@ -36,8 +36,8 @@ typedef void (State::*ActionHandler)(Action*);
 /**
  * Surface that the user can interact with.
  * Specialized version of the standard Surface that processes all the various
- * SDL events and turns them into useful interactions with the Surface, so
- * specialized subclasses don't need to worry about it.
+ * SDL events and turns them into useful interactions with the Surface
+ * so specialized subclasses don't need to worry about it.
  */
 class InteractiveSurface
 	:
@@ -45,7 +45,7 @@ class InteractiveSurface
 {
 
 private:
-	static const int NUM_BUTTONS = 7;
+	static const Uint8 NUM_BUTTONS = 7;
 	static const SDLKey SDLK_ANY;
 
 	Uint8 _buttonsPressed;
